Modify and enhance functionality
was able to add classes to ingrediants and steps to be able to style specific classes without using type selecters and add specificity.

also am trying to add images to home page when hovering over recipe links so users may see a preview of the link they will be clicking on.